requirements.txt
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
Pillow>=10.0
PyYAML>=6.0
scikit-learn>=1.4
seaborn>=0.13
torch>=2.2
torchvision>=0.17

